moving some files from the root renderer folder
