+++ renderer/filmstrip/rating.js
@@ -1,5 +1,5 @@
 const log = require('../../tools/log.js')('filmstrip-rating');
-const exiftool = require('../exiftool-child.js');
+const exiftool = require('../tools/exiftool-child.js');
 
 function ratingControl(rating) {
   function star(setTo) {
